updated 95% - Delete Product - Fixed minor bugs

+++ lib/models/section.dart
@@ -101,7 +101,8 @@ class Section extends ChangeNotifier {
 
     for (final original in originalItems) {
       try {
-        if (!items.contains(original)) {
+        if (!items.contains(original) &&
+            (original.image as String).contains('firebase')) {
           final ref =
               await storage.getReferenceFromUrl(original.image as String);
           await ref.delete();
@@ -121,11 +122,13 @@ class Section extends ChangeNotifier {
   Future<void> delete() async {
     await firestoreRef.delete();
     for (final item in items) {
-      try {
-        final ref = await storage.getReferenceFromUrl(item.image as String);
-        await ref.delete();
-      } catch (e, ex) {
-        debugPrint("$e >>> $ex");
+      if ((item.image as String).contains('firebase')) {
+        try {
+          final ref = await storage.getReferenceFromUrl(item.image as String);
+          await ref.delete();
+        } catch (e, ex) {
+          debugPrint("$e >>> $ex");
+        }
       }
     }
   }
